[move] RefDictをgenericパッケージに移す
